python/requirements.txt
pretrainedmodels==0.7.0
torch==0.4.0a0+200fb22
requests==2.19.1
Pillow==5.2.0
tqdm
